requirements.txt
# Core ML frameworks (Apple Silicon optimized)
mlx>=0.12.0
mlx-lm>=0.12.0

# Hugging Face ecosystem
transformers>=4.36.0
datasets>=2.16.0
huggingface-hub>=0.20.0
tokenizers>=0.15.0
accelerate>=0.25.0

# QLoRA / PEFT
peft>=0.7.0
bitsandbytes>=0.41.0  # For non-MLX fallback

# Training utilities
safetensors>=0.4.0
sentencepiece>=0.1.99
protobuf>=3.20.0

# Evaluation
lm-eval>=0.4.0
rouge-score>=0.1.2
nltk>=3.8.0

# Utilities
numpy>=1.24.0
pandas>=2.0.0
pyyaml>=6.0
tqdm>=4.66.0
wandb>=0.16.0  # Optional: experiment tracking
rich>=13.0.0   # Pretty console output

# Development
pytest>=7.4.0
black>=23.0.0
ruff>=0.1.0
